+++ src/Symfony/Component/Lock/Tests/Store/StoreFactoryTest.php
@@ -16,6 +16,7 @@ use Symfony\Component\Cache\Adapter\AbstractAdapter;
 use Symfony\Component\Cache\Traits\RedisProxy;
 use Symfony\Component\Lock\Store\FlockStore;
 use Symfony\Component\Lock\Store\MemcachedStore;
+use Symfony\Component\Lock\Store\MongoDbStore;
 use Symfony\Component\Lock\Store\PdoStore;
 use Symfony\Component\Lock\Store\RedisStore;
 use Symfony\Component\Lock\Store\SemaphoreStore;
@@ -48,6 +49,10 @@ class StoreFactoryTest extends TestCase
         yield [new \Predis\Client(), RedisStore::class];
         if (class_exists(\Memcached::class)) {
             yield [new \Memcached(), MemcachedStore::class];
+        }
+        if (class_exists(\MongoDB\Collection::class)) {
+            yield [$this->createMock(\MongoDB\Collection::class), MongoDbStore::class];
+            yield ['mongodb://localhost/test?collection=lock', MongoDbStore::class];
         }
         if (class_exists(\Zookeeper::class)) {
             yield [$this->createMock(\Zookeeper::class), ZookeeperStore::class];